refactor PullController
